+++ src/main/java/dev/ftb/mods/ftbic/events/ElectricNetworkLifecycleHandler.java
@@ -2,7 +2,8 @@ package dev.ftb.mods.ftbic.events;
 
 import dev.ftb.mods.ftbic.FTBIC;
 import dev.ftb.mods.ftbic.block.entity.ElectricBlockEntity;
-import net.minecraft.world.level.Level;
+import dev.ftb.mods.ftbic.util.CableRouteCache;
+import net.minecraft.server.level.ServerLevel;
 import net.neoforged.bus.api.SubscribeEvent;
 import net.neoforged.fml.common.EventBusSubscriber;
 import net.neoforged.neoforge.event.level.LevelEvent;
@@ -11,8 +12,9 @@ import net.neoforged.neoforge.event.level.LevelEvent;
 public final class ElectricNetworkLifecycleHandler {
 	@SubscribeEvent
 	public static void onLevelUnload(LevelEvent.Unload event) {
-		if (event.getLevel() instanceof Level l) {
+		if (event.getLevel() instanceof ServerLevel l) {
 			ElectricBlockEntity.forgetElectricNetwork(l);
+			CableRouteCache.forget(l);
 		}
 	}